requirements.txt
Flask==2.1.3
flask-apispec==0.11.1
apispec==5.2.2
gunicorn==23.0.0
marshmallow==3.18.0
SQLAlchemy==1.3.0b3
pymysql==1.1.1
Werkzeug==2.2.0
python-dotenv==1.1.0
cryptography==2.3.1
celery==5.5.0
redis==5.0.4
requests==2.26.0